[UPDATE] Lookupcontainer size and Lookupontainer Componentization
